Changed the `max_num` value from `sys.maxsize` to the number of boxes in the NMS input.
To solve problems with a possible changing in this value when reshaping the model, it was added as an additional input for the NMS.

Results of local performance measurements with using benchmark_app:
|  | Before | After |
| --- |:---:| :---:|
| Count (iterations) | 6 | 170 |
| Duration (ms) | 72126 | 60134 |
| Latency (ms) | 12415 | 351 |
| Throughput (FPS) | 0.08 | 2.84 |